requirements.txt
aiosmtpd
pyyaml
toml
